+++ HISTORY.md
@@ -8,6 +8,7 @@ unreleased
   * remove unnecessary devDependency `safe-buffer`
   * remove `unpipe` package and use native `unpipe()` method
   * remove unnecessary devDependency `readable-stream`
+  * refactor: use object spread to copy error headers
 
 v2.0.0 / 2024-09-02
 ==================